doctime_automation_project/doctime_automation/requirements.txt
Appium-Python-Client==3.1.0
pytest==7.4.0
